set proxy for verde, fix lint issues, remove unused code
