+++ tests/hf_hydrodata/test_data_model_access.py
@@ -11,12 +11,10 @@ import hf_hydrodata.data_model_access
 
 
 def test_load_model():
-    """Test loading of CSV files using load_model."""
+    """Test load_data_model."""
 
-    data_model = hf_hydrodata.data_model_access.load_data_model(False)
+    data_model = hf_hydrodata.data_model_access.load_data_model()
 
-    # NOT FOR SQL assert len(data_model.table_names) >= 14
-    # NOT FOR SQL assert len(data_model.get_table("grid").row_ids) >= 6
     assert data_model.get_table("grid").get_row("conus2").get_value("shape")[1] == 3256
     assert data_model.get_table("grid").get_row("conus2")["shape"][1] == 3256
 
@@ -24,7 +22,7 @@ def test_load_model():
 def test_unit_types():
     """Test loading unit types."""
 
-    data_model = hf_hydrodata.data_model_access.load_data_model(False)
+    data_model = hf_hydrodata.data_model_access.load_data_model()
     units_table = data_model.get_table("units")
     assert units_table.get_row("m3/h")["unit_type"] == "volume_flux"
